fix: session pool should only create session if pending<=waiters
The session pool would always initiate the creation of a new session
if one was not available at the moment that an application requested
a session, even when a large number of sessions were already pending
creation. This could cause the pool to create more sessions than
necessary for an application.

Fixes #790.