fix: retry executeStreamingSql when error code is retryable
The streaming call executeStreamingSql is not automatically retried by gax, as
the gapic configuration for the call does not specify any error codes that should
automatically be retried. Instead, the PartialResultStream is responsible for
retrying these calls with the appropriate resume token. Until now, the call was
only retried when a valid resume token had been seen for the stream, meaning that
if the initial call failed with a retryable error code (e.g. UNAVAILABLE), the
stream would fail with this error. This fix ensures that the call is also retried
when the error occurs for the initial call or before the stream has returned a
valid resume token.

Fixes #620.