fix: pause request stream on backpressure
The request stream should be paused if the downstream is indicating
that it cannot handle any more data at the moment. The request stream
should be resumed once the downstream does accept data again. This
reduces memory consumption and potentially out of memory errors when
a result stream is piped into a slow writer.

Fixes #934