fix: pause request stream on backpressure (#936)
* fix: pause request stream on backpressure

The request stream should be paused if the downstream is indicating
that it cannot handle any more data at the moment. The request stream
should be resumed once the downstream does accept data again. This
reduces memory consumption and potentially out of memory errors when
a result stream is piped into a slow writer.

Fixes #934

* fix: do not retry stream indefinitely

PartialResultSetStream should stop retrying to push data into the
stream after a configurable number of retries have failed.

* fix: process review comments

* fix: remove unused code

* tests: add test for pause/resume

* fix: return after giving up retrying + add test