fix: chained promises to guarantee deletion order
Previously, in the delete code sample, due to the async callbacks given
to `runTransaction`, it was possible for the last delete query to be
executed before the second last delete query. The deletion order is
important so now using `runTransactionAsync` with the promises chained
to guarantee the deletion order.

Fixes #1147