test: skip tests that filter on get operations
Due to a backend bug, if a filter is specified on a
ListDatabaseOperations or ListBackupOperations RPCs, the call fails with
`Error: 3 INVALID_ARGUMENT: Invalid ListDatabaseOperations request.`

The issue is being investigated by the Spanner team. Skipping these
tests in the meantime to unblock merging PRs. #1245 will be used to
track unskipping the tests when a fix is available.

Fixes #1241, #1242, #1243