fix: remove acquire timeout listener on return of session
If an acquire timeout had been set for the session pool, the timeout handler
would not be cleared once a session was successfully returned, or when the
getSession method would fail because of another error. That could cause an
application running for at most acquireTimeout milliseconds if the last action
the application did was to acquire a session.

Fixes #1324