fix: unknown errors should not be retried
Unknown errors that are returned during a transaction should not cause the entire
transaction to be automatically retried, as there is no absolute guarantee that the
initial attempt did not succeed. A `Transaction Outcome Unknown` (or other Unknown)
error could be returned by the `Commit` RPC, and in that case there is a possibility
that the transaction was actually committed.

Fixes #1387