fix: replace projectId placeholder in formatted names
The {{projectId}} placeholder was only replaced in the names of the objects that
were included in the requests, and not in the headers and formatted names. This
would cause the CLOUD_RESOURCE_HEADER to be filled with the {{projectId}} placeholder
instead of the actual value.
As the actual project ID is gotten through an async method, and the constructor and
`instance` methods are synchronous, we can only replace the actual values once it is
retrieved for the first request.

Fixes #1375