feat(default_retention_period): Changes to backup.create API to support the default retention period feature
With the default retention period feature, null value for expiration time is a valid value. If the
expiration time is not specified (null) at the time of create backup then the backup will be
retained for the duration of maximum retention period.

In this commit, expireTime is made optional in CreateBackupOptions and backup.create API
is fixed to handle the null value in the CreateBackupOptions.