feat(default_retention_period): Changes to backup.copy API to support the default retention period feature
With the default retention period feature, null value for expiration time is a valid value.
If the expiration time is not specified (null) at the time of copy backup then the
source backup expiration time will be used for expiring the destination backup as well.

In this commit: expireTime is already optional in CopyBackupOptions and no changes are
required to handle the case of missing expireTime in CopyBackupOptions. Adding unit-tests
for copy backup for the current scenario and a few other scenarios (since they seem to
be missing).