observability: implement trace wrapping, alternative to manual tracing
This change implements trace wrapping by intercepting the
Object.prototype method and starting a span right before invoking
the original method, then an appropriate pass-through replacement
object.

Trace-Wrapping is an alternative to the manual tracing in PR #2087
but then a precursor to use later on using the
opentelemetry-instrumentation framework for which we shall be able
to extract out this code.

Despite being an alternative to manual tracing, in some cases where
perhaps it might be convenient and time saving to use manual tracing,
these two methods can live side by side.

Later on after this change is accepted, we can simply just invoke:

    traceWrap(Class, [methods]);

Updates #2079